+++ src/components/HoldingsRecordFilters/HoldingsRecordFilters.test.js
@@ -22,6 +22,7 @@ jest.mock('../../facetUtils', () => ({
 
 const activeFilters = {
   [FACETS.SHARED]: ['shared1'],
+  [FACETS.HELD_BY]: ['HELD_BY1'],
   [FACETS.EFFECTIVE_LOCATION]: ['loc1'],
   [FACETS.HOLDINGS_PERMANENT_LOCATION]: ['loc2'],
   [FACETS.HOLDINGS_TYPE]: ['loc3'],
@@ -37,6 +38,7 @@ const resources = {
     resource: 'facets',
     records: [{
       'shared': { values: ['shared1'] },
+      'holdings.tenantId': { values: ['heldby1'] },
       'items.effectiveLocationId': { values: ['effectiveLocationId1'] },
       'holdings.permanentLocationId': { values: ['permanentLocationId1'] },
       'holdings.statisticalCodeIds': { values: ['statisticalCodeIds1'] },
@@ -59,6 +61,7 @@ const data = {
   modesOfIssuance: [],
   tagsRecords: [],
   natureOfContentTerms: [],
+  consortiaTenants: [],
   query: [],
   onFetchFacets: jest.fn(),
   parentResources: resources,
@@ -100,12 +103,23 @@ describe('HoldingsRecordFilters', () => {
     });
   });
 
+  it('Should Render Held by, Clear selectedfilters buttons', async () => {
+    renderHoldingsRecordFilters();
+    const heldBy = document.querySelector('[id="accordion-toggle-button-tenantId"]');
+    userEvent.click(heldBy);
+    const Clearselectedfilters = screen.getAllByRole('button');
+    userEvent.click(Clearselectedfilters[3]);
+    await waitFor(() => {
+      expect(onClear).toBeCalledWith(FACETS.HELD_BY);
+    });
+  });
+
   it('Should Render effectiveLocation, Clear selectedfilters buttons', async () => {
     renderHoldingsRecordFilters();
     const effectiveLocation = document.querySelector('[id="accordion-toggle-button-effectiveLocation"]');
     userEvent.click(effectiveLocation);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[3]);
+    userEvent.click(Clearselectedfilters[5]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.EFFECTIVE_LOCATION);
     });
@@ -116,7 +130,7 @@ describe('HoldingsRecordFilters', () => {
     const holdingsPermanentLocation = document.querySelector('[id="accordion-toggle-button-holdingsPermanentLocation"]');
     userEvent.click(holdingsPermanentLocation);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[5]);
+    userEvent.click(Clearselectedfilters[7]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_PERMANENT_LOCATION);
     });
@@ -127,7 +141,7 @@ describe('HoldingsRecordFilters', () => {
     const holdingsType = document.querySelector('[id="accordion-toggle-button-holdingsType"]');
     userEvent.click(holdingsType);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[7]);
+    userEvent.click(Clearselectedfilters[9]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_TYPE);
     });
@@ -138,7 +152,7 @@ describe('HoldingsRecordFilters', () => {
     const holdingsDiscoverySuppress = document.querySelector('[id="accordion-toggle-button-holdingsDiscoverySuppress"]');
     userEvent.click(holdingsDiscoverySuppress);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[9]);
+    userEvent.click(Clearselectedfilters[11]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_DISCOVERY_SUPPRESS);
     });
@@ -149,7 +163,7 @@ describe('HoldingsRecordFilters', () => {
     const holdingsStatisticalCodeIds = document.querySelector('[id="accordion-toggle-button-holdingsStatisticalCodeIds"]');
     userEvent.click(holdingsStatisticalCodeIds);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[11]);
+    userEvent.click(Clearselectedfilters[13]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_STATISTICAL_CODE_IDS);
     });
@@ -160,7 +174,7 @@ describe('HoldingsRecordFilters', () => {
     const holdingsCreatedDate = document.querySelector('[id="accordion-toggle-button-holdingsCreatedDate"]');
     userEvent.click(holdingsCreatedDate);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[13]);
+    userEvent.click(Clearselectedfilters[15]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_CREATED_DATE);
     });
@@ -171,7 +185,7 @@ describe('HoldingsRecordFilters', () => {
     const holdingsUpdatedDate = document.querySelector('[id="accordion-toggle-button-holdingsUpdatedDate"]');
     userEvent.click(holdingsUpdatedDate);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[20]);
+    userEvent.click(Clearselectedfilters[22]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_UPDATED_DATE);
     });
@@ -181,7 +195,7 @@ describe('HoldingsRecordFilters', () => {
     const holdingsSource = document.querySelector('[id="accordion-toggle-button-holdingsSource"]');
     userEvent.click(holdingsSource);
     const Clearselectedfilters = screen.getAllByRole('button');
-    userEvent.click(Clearselectedfilters[27]);
+    userEvent.click(Clearselectedfilters[29]);
     await waitFor(() => {
       expect(onClear).toBeCalledWith(FACETS.HOLDINGS_SOURCE);
     });